refactor: directory cleanup
